Move common modal test methods into modalUtils.js
